Move make_sql_client helper to sql_client_fixtures
This helper function was tagged as only being used in tests, but
it was sitting in the main package. This moves it to the test package
to avoid confusion as we restructure around dbt adapter integrations.
